requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
# Bancos de dados
PyMySQL==1.1.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
# Para configurações
python-dotenv==1.0.0
# Para cache (opcional)
redis==4.6.0